fix order of function calls && bugfix on `cd`

+++ src/builtin/exec.c
@@ -6,13 +6,13 @@
 /*   By: msloot <msloot@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 14:53:33 by msloot            #+#    #+#             */
-/*   Updated: 2024/11/01 15:19:51 by msloot           ###   ########.fr       */
+/*   Updated: 2024/11/01 17:12:18 by msloot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-int	builtin(int argc, char **argv, char **envp)
+int	builtin(size_t argc, char **argv, char **envp)
 {
 	if (ft_strcmp(argv[0], "echo") == 0)
 		return (echo(argc, argv, envp));
